move menu button to own file
