+++ app/src/components/core/menu/index.tsx
@@ -1,1 +1,2 @@
 export * from "./Menu";
+export * from "./MenuButton";
